Update patch set 1
Patch Set 1:

Just to be clear- this is exactly what we did in CM10.  The issue is that the AFE proxy device only supports 48K, and BlueZ/Bluedroid is hardcoded for 44.1. I had "fixed" it in my local tree this way, but had the bad sound quality, which turns out to be Google's bug- not ours. Then QC dropped this patch today, so I guess it's the solution.

Patch-set: 1