Update patch set 1
Patch Set 1:

Hi Ricardo,

thanks for your answer. This is exactly how we want to do it. See BoardConfig.mk[1] and bdroid_buildcfg.h[2]. The problem is that for libbt-hci this config file is not included.
external/bluetooth/bluedroid/Android.mk defines bdroid_CFLAGS and bdroid_C_INCLUDES if BOARD_BLUETOOTH_BDROID_BUILDCFG_INCLUDE_DIR is set. bdroid_CFLAGS is then used in the subfolder Android.mk files for bta, gki, main, stack and utils. But not for hci where the variable we want to define is used.

I hope this clarifies the problem. But maybe I missed something. If you prefer this conversation on IRC I will get in touch with you there. 

[1] https://github.com/CyanogenMod/android_device_htc_endeavoru/blob/cm-10.1/BoardConfig.mk#L141
[2] https://github.com/CyanogenMod/android_device_htc_endeavoru/blob/cm-10.1/bluetooth/bdroid_buildcfg.h

Patch-set: 1