Update patch set 3
Patch Set 3: Code-Review+1

I ageee that it's still not 100% perfect.  I'm still investigating other issues but they seem to be independent of this one and this one, for me, clearly makes the audio immensely better.

PS 2 fixes all of Danny Baumann's suggestions.
PS 3 fixes the brace and a whitespace error I missed, I also updated the commit message to indicate this is an improvements and removed the hope that it was a fix.

Patch-set: 3