Update patch set 1
Patch Set 1: Code-Review+1

I was testing it on several AOSP ROMs (CM11, TwistedKat, SlimKat, CandyKat, Infamous, Beanstalk, PA, PAC) and on all the patch solved the issue. No side effects observed. This issue is device wide not specific and it should have been included since Android 4.2 (when the problem appeared) It'll be great if it'll be permanently merged so it'll be automatically included in any ROM based on CM11

Patch-set: 1