Update patch set 1
Patch Set 1:

Is there a JIRA incident associated with this? When I saw the original commit I was pretty excited since I'm definitely experiencing bad bluetooth audio and I was hoping this would fix it but it was not yet in a nightly so I didn't have a chance to test it yet.  Maybe there could be an option for it on those devices where it helps?

Patch-set: 1
Reviewer: Michael Russo <9660@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0