Update patch set 1
Patch Set 1:

Some history-

Bluedroid's A2DP is very sensitive to changes caused by the kernel scheduler. A common scenario was a PowerHAL which changed something like the timer tick rate when the screen went off. A missed tick or two would spiral out of control in A2DP. This hack was merged which resolved it for many users.

At this point, it is causing harm to my precious Bacon.

Patch-set: 1