Update patch set 1
Patch Set 1:

That's exactly the issue I have (CYAN-4148).  Iker, if you could update the ticket, I'd appreciate it.  Right now it looks like only I have the issue.  I switched to the AOSP branches for bluetooth to try to narrow down the culprit.  I may have found it and it may be the 1.5 commit, but I need to confirm which will probably be tomorrow.

Patch-set: 1