Update patch set 1
Patch Set 1:

That's why I opened the JIRA ticket because it doesn't work consistently.  For me it depends on when I try to initiate audio on the car side (turn on the radio and select bluetooth).  If I let it connect and such when turning on the car and wait up to a minute, it helps.

Patch-set: 1