Update patch set 2
Patch Set 2: Code-Review+1

I saw very few GKI errors on my Civic but when I did it was when I tried to force the skipping with rapid screen/on off actions.  To me it seems better though.  Usually it takes a few days before I see any major skipping.  For me it's better but not resolved.  I vote for it if we think it will overall help BT audio.  We should extend testing beyond m7 though.

Patch-set: 2
Label: Code-Review=+1