Update patch set 1
Patch Set 1: Code-Review+1

The caf value of "2" seems artificially low to me.  I think slayher was seeing problems on some device when it was up around 50-60.

I pushed up this commit so I could fix jittery a2dp on victara but I wonder if it might really just be a better idea to change the value of 2 to 10 or 20 (I'm using 20 on victara).

Patch-set: 1
Label: Code-Review=+1