Update patch set 1
Patch Set 1:

Steve,

I just noticed my link got mangled.  Basically it looks like Google in the aosp branches has the same kind of throttling but it was set at 20 frames and then in October they reduced it to 10 frames to accommodate certain car kits which apparently throttle above this limit.

If this is also affecting bacon (which Andrew just said) then it would be a better idea to abandon this commit and set the default limit to 10 to match what Google is doing.

Patch-set: 1