Update patch set 1
Patch Set 1:

Not sure if this is related but I picked this in my last build (hammerhead) a few days ago and tested with a bluetooth sports watch that usually would disconnect and reconnect multiple times while in use and it seems to have cleared that up. I tested for over an hour and a half with no dropouts yesterday.

Patch-set: 1
Reviewer: Moshe S <10858@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0