Update patch set 5
Patch Set 5:

Nicholas,

If I had to guess, I bet you are building the code with a 44100 sample rate and then setting 48000 in the audio_policy.conf.

I just pushed up two commits (dependent on this one) which should correctly handle that broken configuration and force the output to 44100.

Patch-set: 5