Update patch set 1
Patch Set 1:

I verified what this code is doing on my device (victara == moto x 2014) and it ends up selecting to run 2 iterations that send 5 frames worth of data which actually makes this equivalent to the 10 that I wanted in the first place.

Note that the 5 comes from the fact that I'm connecting to a device that only supports 2mbs mode.  If it was a better device that supported higher speeds, it would already be sending even more data than the non-edr case.

Patch-set: 1