Update patch set 1
Patch Set 1:

Steve,

IF we leave DYN_SAMPLERATE in the code then I think that this set of 3 commits should be merged because they clean things up and make it better.

However, if you merge this commit

http://review.cyanogenmod.org/#/c/88708/

then I really don't see any reason that the dynamic sample rate shouldn't work so it might be worth trying to revert back to always being dynamic and testing that.

If that seems to work, I would still want to see the commit which avoid uninitialized values merged from this set of three commits.

Patch-set: 1