Update patch set 2
Patch Set 2:

I haven't seen an actual instance of this being a problem.

I made this change because I instrumented the code to sometimes return the INCALL failure and when testing that code I found that it sometimes killed the a2dp stream.  This commit fixed that test case.

Patch-set: 2