Update patch set 2
Patch Set 2:

This commit definitely fixes the problems i list in the commit message, but i feel like it's pretty risky to merge.

Unless you think it's a particularly good idea, I would probably recommend -2'ing this.

It may however be worth submitting a new change that reacquires the lock in the error case (that is really just plain broken).

Patch-set: 2