Update patch set 1
Patch Set 1:

That commit (initialize the cfg values) only does something if you have your device misconfigured (built for 48K and configured for 44100).

For me, it fixes a crash.  The fact that the log is spammed with correct values (format = 1 and channels = 3 is right) indicates that this isn't related to the possibility of uninitialized variables.

Patch-set: 1