Update patch set 2
Patch Set 2:

Note that the corresponding AOSP change is identical in the non-EDR case to what we end up with *after* this change.  The only difference was the addition of the truncation of pcm.counter.

The aosp commit is 01d102b3b187ab0d24050c1e12f96e6ee1ce50be

Patch-set: 2