Update patch set 1
Patch Set 1:

In "full disclosure", I'm still having some a2dp issues on victara but it's a new bringup so it may have device specific issues.

However, my gut tells me that I'm fighting with more than one problem.  Right now, I have had audio simply stop playing in the middle of playing a stream and sometimes failing to play after a long pause.

However, after this (and possibly some of my other changes although I think those aren't likely to be fixing any very common issues), I did have a state where I could not play to either a pair of headphones or a speaker and after leaving it disconnected for a few hours (went out), it connected and is playing fine with the headphones.  This is different from the failures I've had most frequently in the past.

Patch-set: 1