Update patch set 2
Patch Set 2: Code-Review+1

If that's the case this needs to be merged asap I wanted to test it in a build myself but haven't been able to make one yet. I've had this (along with the distortion that's now fixed) for quite a while on CM12 and that's one of the few bugs left I have to deal with.

Patch-set: 2