Update patch set 1
Patch Set 1: Code-Review+1

This commit fixes a problem where I lose a2dp output entirely.  This can happen most frequently when starting a stream (the timer gets kicked off by the stream start which increases the odds of a race condition) and it also fixes a problem where the audio suddenly goes silent in the middle of playing a stream.

Patch-set: 1