Update patch set 3
Patch Set 3:

I think that this commit should be merged (despite my earlier concern).

My earlier concern was based on the fact that I get very nervous holding a lock and making a network call (send).  However, we are sending to a local socket and using poll() to ensure that the socket is ready to receive data.

The race condition is definitely real (you can't use a fd that could be concurrently closed as that fd could be reused by another thread before we get a chance to call send).

Patch-set: 3