Update patch set 3
Abandoned

The code should be functionally equivalent with or without this change.  I was only seeing improved a2dp behaviour because this change must have been changing the timing of some other race condition(s).

Since it should be equivalent and this only makes it messier, abandon.

Patch-set: 3
Status: abandoned