Update patch set 3
Patch Set 3: Code-Review-1

I'm not sure that this is necessary or covers enough if it is necessary.

First, all this is wonky because there can only be 1 stream open at a time (a2dp_dev->output is set to NULL in close and set to the stream in open).

I believe that they have this wonkiness because some of the bluetooth state is passed with the device and some via the stream (the adev_set_parameters passes the parameters to the output stream and they are important parameters).

Because of that we don't have enough locking if we are protecting all concurrent access.

If we can have concurrent open and close then all this is scary because the close to set a2dp_dev->output to NULL even if another stream has been concurrently opened.  If anything, that should be an atomic test and set and the free(stream) should be moved below the atomic test and set (so that same pointer can't be reused by the new open).

Alternatively, we could lock the open_output_stream and make the locking wider in the close to deal with that.

Finally, on to why I think that this isn't particularly useful.

If there is any concurrent access of the stream while the close is being called, the whole thing is likely to crash and burn both with and without the lock.

Immediately after the lock we free(stream) and therefore even the lock itself that people would be blocked on is being freed and could crash.  That is, I'm not sure that blocking access to a structure that we immediately free is particularly useful.

Patch-set: 3