Update patch set 1
Patch Set 1: Code-Review-1

The code is:

if working in edr mode
   A
else
   B

and in aosp the code is just "B" without the if/else.  In B, they reduced the number of frames sent from 20 to 10 because 20 caused problems with specific cars.

On victara, I added some extra logging to see that at 2 mbps it would send 5 frames per iteration and at most 2 iterations in the "A" case (aka the same 10 as in "B").  This seems like a good limit to me if sending too much data causes problems for some carkits.

Patch-set: 1