Update patch set 1
Patch Set 1:

I just added some new log messages where it determined if the device was edr and when it determines the maximum number of frames that can be sent without violating the mtu of the device and finally looked to see how many frames was actually being sent to see it was doing 5x2 (max) per tick.

I don't think that there is existing logging that you can just magically enable to see it.

Patch-set: 1