Update patch set 1
Patch Set 1: Code-Review+1

I think that this is good to go.  On many (most) devices (aka all the ones that had problems before), the timer can go back in time and that causes bad things to happen.

I'm not really sure what bad things happen in this code is the timer goes back in time (it will think that a huge amount of data is waiting to be sent and god only knows what it will send and when).

Either way, agreeing within the code on a standard "now_us" is a good thing.

Patch-set: 1
Label: Code-Review=+1