Update patch set 2
Patch Set 2:

FYI, this is working great.  I've been listening to hours of music and haven't noticed a single audio glitch.

This, obviously, needs a lot more testing than me and one device.

Prior to this commit, I was seeing audio glitches when the timer was slow to fire (>= 120ms was almost always a glitch).

I think that the fundamental problem is due to the rounding down, we are always riding very close to the level at which the sink starts to run out of data and the longer timer interval then causes it to starve.

I removed the Audio Congestion logging.  I think this is due to the fact that we are (correctly) figuring out that there is a lot more data that could be sent it shows a lot of congestion but never has a problem (even with the extra log spam) keeping the buffers full.

Patch-set: 2
Reviewer: Christopher R. Palmer <6351@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0