Update patch set 2
Patch Set 2:

Yeah, I know it doesn't make sense that this could cause an issue, but I recompiled bluetooth.default.so without this commit as the only change and I didn't have this issue.  Also, the audio corruption/speedup only seems to happen with this commit present.  Further testing is needed, but right now this is the only lead I have for this issue.

Patch-set: 2