Update patch set 1
Patch Set 1:

Matt,

What exactly is the "klte stuttering problem"?  Right now for me and my devices, I'm working toward the last 1% of playback where it is normally good but every now and again I get a slight drop-out / skip.

Right now, I think that the only real problem I have is every now and again a get a slight drop-out when the 20ms tick takes over about 120ms to get triggered and otherwise playback seems to be good.

But, as I said, this could be firmware dependent (or kernel, or device config).

Patch-set: 1