Update patch set 2
Patch Set 2:

The commit message usually indicates what it fixes, so it should be pretty self explainable. The issues that they fix might seem small and pointless, and although most of us probably won't run into them, that doesn't mean someone else won't. Specifically, on my device, there's been an issue with that when turning off bluetooth, it sometimes goes into a loop of turning itself on and off. This particular patch seems to fix that.

Patch-set: 2