Update patch set 2
Patch Set 2:

CAF's track record in bluetooth doesn't inspire me.  If it is actually fixing something that you can see then it's probably fine to merge it.  I would hate to merge things that aren't fixing a problem one of us can reproduce and/or for which we don't really understand the fix.

Patch-set: 2