Update patch set 2
Patch Set 2:

We're getting reports on this from Oneplus One users.

If we set sampling_rates to any of:

44100
48000
44100|48000

in audio_policy.conf, the behavior is still triggered.  The specific 'symptom' is the audio playback is sped up, followed by short gaps.  It seems like a 44.1khz sample is being handled by the radios as a 48khz sample.

Patch-set: 2