Update patch set 2
Patch Set 2:

If user feedback is valued, I've had numerous complaints about no audio heard and absolutely horrible audio quality on car units and on bluetooth headsets. Both on CM 12.0 and 12.1. CM11 is apparently fine, so many users revert to either that or stock rom because of the bluetooth.

Patch-set: 2