Update patch set 2
Patch Set 2:

Pat, are you able to reproduce the problem yourself and see that this reverts fixes it?

If so, it would be worth investigating it more.  You can instrument the code to log the pcm_sample_rate in btif/src/btif_media_task.c and see if a bad value is being used or if it is always using the value that you've configured in the audio_policy.conf (pick either 44100 or 48000, whichever is the right value for that phone).

FYI, I don't have any devices that exhibit any problems on my phone (I have headphones, headset, speaker and car that I use regularly).

Patch-set: 2