Update patch set 2
Patch Set 2:

(1 comment)

No objections to the functionality of this change, but I'd prefer to see the way the buffer is calculated done differently.

Making it clear that we're finding how many ticks we need for the buffer, then finding the bytes for those ticks, then the number of frames that the buffer ends up being will help a lot for the future maintainability of this.

Patch-set: 2