Update patch set 4
Patch Set 4:

Jeremy, if you cherry-picked this commit then it should not have changed anything from the revert.

I was mostly testing without that revert because I am 99.9999% sure that doesn't change anything with my phone and I didn't want to introduce another variable in my testing (I was trying to fix an independent problem).

If, instead of cherry-picking the commit, you checked it out then the latest PS should include the revert.  Please test either a checkout of that PS (even including the next commit too if you want) or a cherry-pick.

Patch-set: 4