Update patch set 2
Patch Set 2:

I don't really know whether this makes anything better.  It seems like it should (the logic in btif_media is to try to keep a perfect balance between the amount of data that we send and the time that has elapsed which sounds like it should risk starving the connection).

However I suspect that most devices pre-buffer themselves so this may really be irrelevant or at best cause a 50ms reduction in the time between playback starting on the phone and playback actually starting on the bt sink.

Patch-set: 2