Update patch set 1
Patch Set 1:

Going back to a more "sane" kernel config I get the wiimote plus working. The lights do keep flashing tho, and it doesn't support any of the extensions like the "snes controller". I think this is the core of the remaining issue: the wiimote extensions HOTPLUG not working. This explains all the issues remaining, LEDS, plug-in addons not working, and potentially even the Wii U Pro Controller not working (in this linux kernel it's actually implemented as an extension). What is odd about the last thing tho, is that it should be able to pair even before any of the hid driver specifics fail, right?

Now to get the hotplug to work...

More info here:
http://git.kernel.org/cgit/linux/kernel/git/torvalds/linux.git/tree/drivers/hid/hid-wiimote.c
http://git.kernel.org/cgit/linux/kernel/git/torvalds/linux.git/tree/drivers/hid/hid-wiimote-modules.c

Patch-set: 1