Update patch set 1
Patch Set 1:

Issue with the wiimote plus seems to be kernel related, more info here

https://github.com/dvdhrm/xwiimote/issues/31
http://git.kernel.org/cgit/linux/kernel/git/torvalds/linux.git/commit/?id=e73dcfbf061b524fe9aaef56cf3c2e234a45ec19

Remains the issue of the Wii U Pro Controller not pairing, which is quite odd. Checking in bluez autopair code, this is the same pin "gimmick" as for all the other ones. And on my desktop the Wii U Pro controller pairs perfectly fine with bluez 5.31.

Patch-set: 1
Label: Code-Review=0