Update patch set 1
Patch Set 1:

Paul, i'm not sure what kernel your workign with but all the devices i've known wiimotes to be tested on have used the 40 or so commits like this htc 8960 kernel. That BT commit looks like it may be relevent, if it helps let us know i may add it to my kernels as well.

https://github.com/CyanogenMod/android_kernel_htc_msm8960/commits/cm-12.1/drivers/hid?page=1

Patch-set: 1