Update patch set 3
Patch Set 3:

I thought I was affected by this, but my issue is different.
I still cannot connect with this patch, but stock rom's work.
I just tried a Zenfone2 running 5.0 stock and it works as well,
so it's something to do with CM, but I don't know what exactly.

Patch-set: 3